MBU Red Gems (#2)
* MBU Red Gems

In the original Marble Blast Ultra, all the gems are colored red every time you play a level. It is not the case in PQ. Having all the gems be red is the icing on the cake needed to make the PQ versions of these levels faithful to the originals.